requirements.txt
faiss-cpu==1.12.0
sentence-transformers>=3.0.0
anthropic>=0.91.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests
beautifulsoup4
lxml
